feat 新增修改信息以及修改密码

+++ django_hrm/models/business.py
@@ -8,10 +8,12 @@
 """
 import logging
 
+from django.contrib.auth import authenticate
 from django.contrib.auth.models import User
 from django.db import models, transaction
 
 from django_test.db.models import BaseModel, UnsignedBigAutoField, UnsignedIntegerField
+from django_test.exceptions import UserPasswordError
 
 
 class Business(BaseModel):
@@ -38,18 +40,22 @@ class Business(BaseModel):
                 business_obj = cls.objects.create(name=user_name)
 
                 user = User.objects.create_user(username=user_name, password=password)
-                user.userprofile.type = UserProfile.UserType.CUSTOMER
+                user.userprofile.type = UserProfile.UserType.BUSINESS
                 user.userprofile.name = user_name
                 user.userprofile.business_id = business_obj.id
                 business_obj.create_user_id = user.userprofile.id
                 business_obj.update_user_id = user.userprofile.id
                 business_obj.save()
                 user.userprofile.save()
-        logging.info('登录顾客用户id为', user.id)
-        logging.info('user_profile的id为', user.userprofile.id)
-        # 生成token
-        token_obj, _ = ApiToken.objects.get_or_create(user=user)
-        if token_obj.is_expired():
-            token_obj.delete()
+        user = authenticate(username=user_name, password=password)
+        if user:
+            logging.info('登录顾客用户id为', user.id)
+            logging.info('user_profile的id为', user.userprofile.id)
+            # 生成token
             token_obj, _ = ApiToken.objects.get_or_create(user=user)
-        return token_obj, user
+            if token_obj.is_expired():
+                token_obj.delete()
+                token_obj, _ = ApiToken.objects.get_or_create(user=user)
+            return token_obj, user
+        else:
+            raise UserPasswordError()
